Added `offset` option to `FlutterMapState.move` and related methods (resolves #1460, #777, #952) Major refactoring and re-organization to improve understandibility
