+++ lib/src/gestures/gestures.dart
@@ -5,7 +5,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/gestures/latlng_tween.dart';
-import 'package:flutter_map/src/map/flutter_map_state.dart';
+import 'package:flutter_map/src/map/state.dart';
 import 'package:latlong2/latlong.dart';
 
 abstract class MapGestureMixin extends State<FlutterMap>
