+++ lib/src/layer/tile_layer/tile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_map/src/core/point.dart';
+import 'package:flutter_map/src/misc/point.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_builder.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_image.dart';
