+++ lib/src/layer/tile_layer/tile_bounds/tile_bounds.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_map/src/core/bounds.dart';
-import 'package:flutter_map/src/geo/crs/crs.dart';
+import 'package:flutter_map/src/misc/private/bounds.dart';
+import 'package:flutter_map/src/geo/crs.dart';
 import 'package:flutter_map/src/geo/latlng_bounds.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_bounds/tile_bounds_at_zoom.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_range.dart';
