+++ lib/src/layer/tile_layer/tile_bounds/tile_bounds_at_zoom.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_map/src/core/point.dart';
+import 'package:flutter_map/src/misc/point.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_coordinates.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_range.dart';
