+++ lib/src/layer/tile_layer/tile_range.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math;
 
-import 'package:flutter_map/src/core/bounds.dart';
-import 'package:flutter_map/src/core/point.dart';
+import 'package:flutter_map/src/misc/private/bounds.dart';
+import 'package:flutter_map/src/misc/point.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_coordinates.dart';
 
 abstract class TileRange {
