+++ lib/src/layer/tile_layer/tile_update_transformer.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:flutter_map/src/core/util.dart';
+import 'package:flutter_map/src/misc/private/util.dart';
 import 'package:flutter_map/src/gestures/map_events.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_layer.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_update_event.dart';
