+++ lib/src/misc/move_and_rotate_result.dart
@@ -1,0 +1,6 @@
+class MoveAndRotateResult {
+  final bool moveSuccess;
+  final bool rotateSuccess;
+
+  MoveAndRotateResult(this.moveSuccess, this.rotateSuccess);
+}
