+++ lib/src/misc/point.dart
