+++ lib/src/misc/private/positioned_tap_detector_2.dart
@@ -1,3 +1,9 @@
+//////////////////////////////////////////////////////////////////
+///              Based on the work by Ali Raghebi              ///
+///           Now maintained here due to abandonment           ///
+/// https://github.com/arsamme/flutter-positioned-tap-detector ///
+//////////////////////////////////////////////////////////////////
+
 import 'dart:async';
 import 'dart:math';
