+++ test/layer/tile_layer/tile_range_test.dart
@@ -1,6 +1,4 @@
-import 'package:flutter_map/src/core/bounds.dart';
-import 'package:flutter_map/src/core/point.dart';
-import 'package:flutter_map/src/layer/tile_layer/tile_coordinates.dart';
+import 'package:flutter_map/plugin_api.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_range.dart';
 import 'package:test/test.dart';
